requirements.txt
twitchAPI
python-telegram-bot
humanize